+++ src/Store.js
@@ -1,3 +1,4 @@
+const collect = require('collect.js');
 const errors = require('./errors.js');
 const Record = require('./Record.js');
 
@@ -9,24 +10,32 @@ const Record = require('./Record.js');
  */
 function Store(records) {
   validateRecords(records);
-  this.records = records;
+  this.collection = collect(records);
 
-  this.all = require('./Store/all.js');
-  this.cast = require('./Store/cast.js');
-  this.chunk = require('./Store/chunk.js');
-  this.count = require('./Store/count.js');
-  this.each = require('./Store/each.js');
-  this.get = require('./Store/get.js');
-  this.map = require('./Store/map.js');
-  this.sortBy = require('./Store/sortBy.js');
-  this.sortByDesc = require('./Store/sortByDesc.js');
-  this.where = require('./Store/where.js');
-  this.dump = require('./Store/dump.js');
-  this.dd = require('./Store/dd.js');
+  this.all         =            this.collection.all.bind(this.collection);
+  this.chunk       = (size)  => this.collection.chunk(size).toArray();
+  this.count       =            this.collection.count.bind(this.collection);
+  this.dd          =            this.collection.dd.bind(this.collection);
+  this.dump        = ()      => console.log(this.collection.all()); // Normalizes dump and dd.
+  this.each        =            this.collection.each.bind(this.collection);
+  this.filter      = (fn)    => new this.constructor(this.collection.filter(fn).all());
+  this.get         = (key)   => this.collection.firstWhere('key', key);
+  this.isEmpty     =            this.collection.isEmpty.bind(this.collection);
+  this.keyed       = ()      => this.collection.keyBy('key').all();
+  this.map         = (fn)    => this.collection.map(fn).all();
+  this.reverse     = ()      => new this.constructor(this.collection.reverse().all());
+  this.sort        = (fn)    => new this.constructor(this.collection.sort(fn).all());
+  this.sortBy      = (mixed) => new this.constructor(this.collection.sortBy(mixed).all());
+  this.sortByDesc  = (mixed) => new this.constructor(this.collection.sortByDesc(mixed).all());
+  this.take        = (size)  => this.collection.take(size).all();
 
-  this.linkToOne = require('./Store/linkToOne.js');
-  this.linkToMany = require('./Store/linkToMany.js');
-  this.linkFromOne = require('./Store/linkFromOne.js');
+  this.where       = require('./Store/where.js');
+  this.pluck       = require('./Store/pluck.js');
+  this.pluckUnique = require('./Store/pluckUnique.js');
+
+  this.linkToOne    = require('./Store/linkToOne.js');
+  this.linkToMany   = require('./Store/linkToMany.js');
+  this.linkFromOne  = require('./Store/linkFromOne.js');
   this.linkFromMany = require('./Store/linkFromMany.js');
 }