+++ tests/blego.Store.reverse.test.js
@@ -1,4 +1,4 @@
-describe('blego.Store.get', () => {
+describe('blego.Store.reverse', () => {
   const Blego = require('Blego.js');
   let blego;
 
@@ -7,19 +7,17 @@ describe('blego.Store.get', () => {
     blego = new Blego();
   });
 
-  it('Gets the record with specified key', () => {
+  it('Reverts the order of Records', () => {
     const store = new blego.Store([
       new blego.Record('1', {}),
       new blego.Record('2', {}),
       new blego.Record('3', {}),
     ]);
 
-    expect(store.get('2').key).toEqual('2');
-  });
+    const reversed = store.reverse().all();
 
-  it('Gets undefined if the key does not exist', () => {
-    const store = new blego.Store([]);
-
-    expect(store.get('2')).toEqual(null);
+    expect(reversed[0].key).toEqual('3');
+    expect(reversed[1].key).toEqual('2');
+    expect(reversed[2].key).toEqual('1');
   });
 });