+++ tests/blego.Store.take.test.js
@@ -1,4 +1,4 @@
-describe('blego.Store.cast', () => {
+describe('blego.Store.take', () => {
   const Blego = require('Blego.js');
   let blego;
 
@@ -7,13 +7,15 @@ describe('blego.Store.cast', () => {
     blego = new Blego();
   });
 
-  it('Gets an object with Record keys as property names', () => {
+  it('Gets an array of specified size from the front of the Store', () => {
     const store = new blego.Store([
       new blego.Record('1', {}),
       new blego.Record('2', {}),
       new blego.Record('3', {}),
     ]);
 
-    expect(Object.keys(store.cast())).toEqual(['1', '2', '3']);
+    const records = store.take(2);
+
+    expect(records.length).toEqual(2);
   });
 });